Authenticate with Travis CI

Authenticates with Travis using your Github account. Returns an access token. The token will be obtained only once in each R session, but it will never be cached on the file system. In most scenarios, these functions will be called implicitly by other functions. auth_travis() only performs the authentication with Travis CI. travis_token() authenticates and checks if the repository is known to Travis CI. If not, a GitHub sync via travis_sync() is performed.

auth_travis(gh_token = NULL)

travis_token(repo = NULL)

Arguments

gh_token
[Token2.0] A GitHub token, by default obtained from auth_github() using the "read:org", "user:email", "repo_deployment", "repo:status", "read:repo_hook", and "write:repo_hook" scopes.

repo
[string] The GitHub repo slug in the format "<user|org>/<repo>", if NULL no sync is performed.