Create a repository on GitHub

Creates a GitHub repository. github_create_repo() only creates the repository on GitHub. use_github() creates a GitHub repository for an existing Git repository, and connects both. new_github() creates an empty Git repository with a single commit and a corresponding GitHub repository, and always pushes to GitHub.

github_create_repo(name, org = NULL, private = FALSE, gh_token = NULL,
  quiet = FALSE)

use_github(path = ".", push = NA, name = NULL, org = NULL,
  private = FALSE, gh_token = NULL, quiet = FALSE)

new_github(path, user = getOption("devtools.desc.author"), name = NULL,
  org = NULL, private = FALSE, gh_token = NULL, quiet = FALSE)

Arguments

name
[string] The name of the new repository on GitHub, default: basename of the repo directory.

org
[string] The organization of the new repository on GitHub, default: user namespace, no organization.

private
[flag] Must be FALSE, creation of private repositories not yet supported.

gh_token
[Token2.0] GitHub authentication token, by default obtained from auth_github() with the "public_repo" and (if an organization repo) "write:org" scopes.

quiet
[flag] Set to FALSE to suppress success message.

path
[string] The path to the existing Git repository.

push
[flag] Should the contents of the existing repository be pushed to GitHub? Defaults to FALSE for non-interactive sessions, and asks the user in interactive sessions.

user
[person|string] An object of class person, or a string that returns such an object when evaluated.

See also

Other GitHub functions: github_add_key, github_create_pat, github_info