Github information

Retrieves metadata about a Git repository from GitHub.

github_info() returns a list as obtained from the GET "/repos/:repo" API.

github_repo() returns the true repository name as string.

uses_github() returns a flag that indicates if this repo is GitHub-enabled. If TRUE, the "info" and "repo" attributes contain the results of the corresponding functions. If FALSE, the "reason" attribute explains what happened during detection of GitHub status.

github_info(path = ".", gh_token = NULL)

github_repo(path = ".", info = github_info(path))

uses_github(path = ".")

Arguments

path
[string]
The path to a GitHub-enabled Git repository (or a subdirectory thereof).

gh_token
[Token2.0]
GitHub authentication token, by default obtained from auth_github() with empty scope.

info
[list]
GitHub information for the repository, by default obtained through github_info().

See also