Getting tic testing up and running

Prepares a repo for building and deploying supported by tic.

use_tic(path = ".", quiet = FALSE)

Arguments

path
[string]
The path to the repo to prepare.

quiet
[flag]
Less verbose output? Default: FALSE.

Details

The preparation consists of the following steps:

If necessary, create a GitHub repository via use_github()

Enable Travis via travis_enable()

Create a default .travis.yml file (overwrite after confirmation in interactive mode only)

Create a default appveyor.yml file (depending on repo type, overwrite after confirmation in interactive mode only)

Create a default tic.R file depending on the repo type (package, website, bookdown, ...)

Enable deployment (if necessary, depending on repo type) via use_travis_deploy()

Create a GitHub PAT and install it on Travis CI via travis_set_pat()

This function is aimed at supporting the most common use cases. Users who require more control are advised to manually call the individual functions.