Travis CI builds

Functions around completed or pending Travis CI builds and jobs.

travis_get_builds() calls the "builds" API for the current repository.

travis_restart_build() restarts a build with a given build ID.

travis_restart_last_build() restarts the last build.

travis_cancel_build() cancels a build with a given build ID.

travis_get_jobs() calls the "jobs" API for the current repository.

travis_restart_job() restarts a job with a given job ID.

travis_cancel_job() cancels a job with a given job ID.

travis_debug_job() restarts, in debug mode, a job with a given job ID. See the Travis CI documentation for more details.

travis_job_log() returns a build job log.

travis_delete_log() deletes a build job log.

travis_get_builds(repo = github_repo(), token = travis_token(repo))

travis_restart_build(build_id, repo = github_repo(),
  token = travis_token(repo), quiet = FALSE)

travis_restart_last_build(repo = github_repo(), token = travis_token(repo),
  quiet = FALSE)

travis_cancel_build(build_id, repo = github_repo(),
  token = travis_token(repo), quiet = FALSE)

travis_get_jobs(build_id, repo = github_repo(), token = travis_token(repo))

travis_restart_job(job_id, repo = github_repo(), token = travis_token(repo),
  quiet = FALSE)

travis_cancel_job(job_id, repo = github_repo(), token = travis_token(repo),
  quiet = FALSE)

travis_debug_job(job_id, log_output = FALSE, repo = github_repo(),
  token = travis_token(repo), quiet = FALSE)

travis_get_log(job_id, repo = github_repo(), token = travis_token(repo),
  quiet = FALSE)

travis_delete_log(job_id, repo = github_repo(), token = travis_token(repo),
  quiet = FALSE)

Arguments

repo
[string|numeric]
The GitHub repo slug, by default obtained through github_repo(). Alternatively, the Travis CI repo ID, e.g. obtained through travis_repo_id().

token
[Token2.0]
A Travis CI token obtained from travis_token() or auth_travis().

build_id
[integer(1)]
The build ID, as obtained from travis_get_builds().

quiet
[flag]
Set to FALSE to suppress success message.

job_id
[integer(1)]
The job ID, as obtained from travis_get_builds().

log_output
Show the debugging output in the publicly visible log? When set to TRUE, refrain from issuing commands that might expose secrets.